Fix incorrect package names in marker tests
