+++ marker-tests/src/test/kotlin/io/spine/core/jvm/marker/EveryIsOptionRendererSpec.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spne.core.jvm.marker
+package io.spine.core.jvm.marker
 
 import com.google.protobuf.Message
 import io.kotest.matchers.shouldBe
